Automated inactivity triggered EKS/ELB clean up for sandboxes.
There is a CloudWatch alarm that will fire after a day of a sandbox
ELB target groups having 0 targets. The alarm will trigger the deletion
of the EKS cluster and ELB load balancers.

To reboot the resources, one needs to first run `terragrunt apply
-target=module.eks_cluster` to recreate the EKS control plane.
Afterwards, run the normal `terragrunt apply`s to recreate the rest of
the resources along with the ELBs.

To opt out of this clean up, use the variables
`disable_inactivity_cleanup`. Use this sparingly only when you need to
save the state for a specific period of time, not as a default.