Automated inactivity triggered EKS/ELB clean up for sandboxes. (#791)
Creates a CloudWatch alarm that aims to detect inactivity. The alarm will trigger the deletion of the EKS cluster and ELB load balancers. It will fire after a workday (excludes weekends) where sandbox ELB target groups have 0 targets. CloudWatch limits the evaluation period to 1 day so this was the best workaround to avoid triggering the inactivity clean-up due to normal activity during weekends.

To reboot the resources, one needs to first run terragrunt apply -target=module.eks_cluster to recreate the EKS control plane. Afterwards, run the normal terragrunt applys to recreate the rest of the resources along with the ELBs.

To opt out of this clean up, use the variables
disable_inactivity_cleanup. Use this sparingly, not as a default.