Test case for Gateway API HTTPRoute in Traefik (#2074)
* feat: add test-feature.bin to .gitignore

* refactor: streamline Traefik ingress route tests and add cleanup function

* feat: enhance DeployTestcase function to accept testName parameter for unique resource naming

* refactor: simplify TestTraefikGatewayHTTPRoute by removing redundant logs and waits

* refactor: remove unnecessary wait in TestTraefikGatewayHTTPRoute

* Add generated registration file for Gateway API v1

This commit introduces a new generated registration file for the Gateway API v1. The file includes the necessary imports, group and version definitions, and the registration of known types such as BackendTLSPolicy, GRPCRoute, Gateway, and HTTPRoute. This change is essential for enabling the proper registration and serialization of these API objects within the Kubernetes ecosystem.

* refactor: update Gateway API scheme registration method in TestTraefikGatewayHTTPRoute